fixed duck mode to simply flip the duck image and offset it as required when flipped.
I honestly don't know what's better, two duck textures and switch to flip, or including the math library to find the absolute values of floats in order to always have the absolute value of sn and using sn to flip the duck and then an ADDITIONAL if statement that needs to be checked each loop just to make sure the offset works. Maybe I shouldn't include duckmode in this program and make a spinoff of BounceWii just to see a duck floating around and maybe make it have two ducks and figure out hitboxes and object orientation. Know what. Screw it. Imma remove duckmode from this bouncewii. It adds too much checking and everything and I plan to go further with the duckmode that bouncewii DEFINITELy doesnt need.